+++ CHANGELOG.rst
@@ -21,6 +21,7 @@ Changed
     - Capitalize event description in UI (#1522)
     - Display event name in UI friendly format (#1524)
     - Display search results with new layout (#1521)
+    - Enable search for display formatting of event name (#1525)
 
 
 v1.0.3 (2024-12-12)